res updated Opcode to version 1.3.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@17996 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/collide/opcode/OPC_PlanesAABBOverlap.h
@@ -24,7 +24,7 @@ inline_ BOOL PlanesCollider::PlanesAABBOverlap(const Point& center, const Point&
 	// variant here uses 3 fabs ops, 6 muls, 7 adds and two floating point
 	// comparisons per plane. The routine early-exits if the AABB is found
 	// to be outside any of the planes. The loop also constructs a new output
-	// clip mask. Most FPUs have a native single-cycle fabs() operation.
+	// clip mask. Most FPUs have a native single-cycle fabsf() operation.
 
 	udword Mask				= 1;			// current mask index (1,2,4,8,..)
 	udword TmpOutClipMask	= 0;			// initialize output clip mask into empty. 
@@ -33,7 +33,7 @@ inline_ BOOL PlanesCollider::PlanesAABBOverlap(const Point& center, const Point&
 	{
 		if(in_clip_mask & Mask)				// if clip plane is active, process it..
 		{               
-			float NP = extents.x*(float)fabs(p->n.x) + extents.y*(float)fabs(p->n.y) + extents.z*(float)fabs(p->n.z);	// ### fabs could be precomputed
+			float NP = extents.x*fabsf(p->n.x) + extents.y*fabsf(p->n.y) + extents.z*fabsf(p->n.z);	// ### fabsf could be precomputed
 			float MP = center.x*p->n.x + center.y*p->n.y + center.z*p->n.z + p->d;
 
 			if(NP < MP)						// near vertex behind the clip plane...
